moved method to new TestUtils
