+++ src/test/java/nl/knaw/dans/layerstore/AbstractTestWithTestDir.java
@@ -18,12 +18,8 @@ package nl.knaw.dans.layerstore;
 import org.apache.commons.io.FileUtils;
 import org.junit.jupiter.api.BeforeEach;
 
-import java.io.ByteArrayInputStream;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
-
-import static org.junit.jupiter.api.Assumptions.assumeTrue;
 
 /**
  * A test class that creates a test directory for each test method.
@@ -44,14 +40,6 @@ public abstract class AbstractTestWithTestDir {
         }
     }
 
-    public static ByteArrayInputStream toInputStream(String testContent) {
-        return new ByteArrayInputStream(toBytes(testContent));
-    }
-
-    public static byte[] toBytes(String testContent) {
-        return testContent.getBytes(StandardCharsets.UTF_8);
-    }
-
     public void createEmptyStagingDirFiles(String... paths) {
         for (String path : paths) {
             var message = "Could not create staging file: " + path;
@@ -65,14 +53,5 @@ public abstract class AbstractTestWithTestDir {
                 throw new RuntimeException(message, e);
             }
         }
-    }
-
-    /**
-     * Assume that a bug is not yet fixed. This allows to skip assertions while still showing the code covered by the test.
-        *
-        * @param message the message to display
-        */
-    public void assumeNotYetFixed (String message) {
-        assumeTrue(false, message);
     }
 }
