+++ src/test/java/nl/knaw/dans/layerstore/LayeredItemStoreListDirectoryTest.java
@@ -20,6 +20,7 @@ import org.junit.jupiter.api.Test;
 
 import java.nio.file.Files;
 
+import static nl.knaw.dans.layerstore.TestUtils.toInputStream;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
 public class LayeredItemStoreListDirectoryTest extends AbstractLayerDatabaseTest {
