+++ src/test/java/nl/knaw/dans/layerstore/ZipArchiveUnarchiveToTest.java
@@ -23,6 +23,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
+import static nl.knaw.dans.layerstore.TestUtils.assumeNotYetFixed;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
